YAY-114 fix: NoAgeNotAllowed, Recaptcha, UploadFileRespons, Post
