+++ src/entities/post/model/index.ts
@@ -1,3 +1,4 @@
 export { type LastPostProps, lastPostsSchema } from './schemas'
 export { publicPostsSchema } from './schemas'
 export { privatePostsSchema } from './schemas'
+export { type PostByIdType } from './types'
