+++ src/entities/post/model/types/index.ts
@@ -1,0 +1,1 @@
+export { type PostByIdType } from './post.types'
