+++ src/entities/post/ui/Post/Post.stories.tsx
@@ -1,8 +1,9 @@
 import { Meta, StoryObj } from '@storybook/nextjs-vite'
 import { Post } from './Post'
-import type { PostByIdType } from '@/features/post/api'
+
 import { Provider } from 'react-redux'
 import { store } from '@/app/store'
+import { PostByIdType } from '@/entities/post/model'
 
 const meta = {
    title: 'Post',
