+++ src/entities/post/ui/Post/PostClientWrapper/PostClientWrapper.tsx
@@ -1,5 +1,5 @@
 'use client'
-import type { PostByIdType } from '@/features/post/api'
+import { PostByIdType } from '@/entities/post/model'
 import { Post } from '@/entities/post/ui/Post'
 import { usePostController } from '@/features/post/model/hooks'
 import { EditPostModal } from '@/features/post-creator/ui/modals/EditPostModal'
