+++ src/features/auth/model/index.ts
@@ -1,2 +1,1 @@
-export * from './NoAgeNotAllowed'
 export * from './authStorage'
