+++ src/features/auth/ui/Recaptcha/index.ts
@@ -1,0 +1,1 @@
+export { Recaptcha } from './Recaptcha'
