+++ src/shared/api/index.ts
@@ -1,0 +1,2 @@
+export { type UploadFileItem } from './types'
+export { type UploadFileResponse } from './types'
