+++ src/shared/api/types/file.types.ts
@@ -1,0 +1,8 @@
+export type UploadFileItem = {
+   url: string
+   fileId: string
+}
+
+export type UploadFileResponse = {
+   data: UploadFileItem[]
+}
