+++ src/shared/api/types/index.ts
@@ -1,0 +1,2 @@
+export { type UploadFileItem } from './file.types'
+export { type UploadFileResponse } from './file.types'
