--- src/shared/ui/Recaptcha/index.ts
@@ -1,1 +1,0 @@
-export * from './Recaptcha'
